moved calculateTimeLeft to helpers
