+++ frontend/app/src/helpers/helpers.ts
@@ -87,3 +87,18 @@ export function sendToRedirect(url) {
   el.target = '_blank';
   el.click();
 }
+
+  export const calculateTimeLeft = (timeLimit: Date): {minutes: number, seconds: number} => {
+    const difference = new Date(timeLimit).getTime() - new Date().getTime();
+
+    let timeLeft: any = {};
+
+    if (difference > 0) {
+      timeLeft = {
+        minutes: Math.floor((difference / 1000 / 60) % 60),
+        seconds: Math.floor((difference / 1000) % 60),
+    };
+  }
+
+  return timeLeft;
+};
