+++ frontend/app/src/people/widgetViews/summaries/wantedSummaries/invoice.tsx
@@ -2,32 +2,17 @@ import React, { useEffect, useState } from 'react';
 import { CopyInvoiceBtn, CountDownText, CountDownTimer, CountDownTimerWrap, InvoiceWrap } from './style';
 import { useStores } from '../../../../store';
 import QR from 'people/utils/QR';
+import { calculateTimeLeft } from '../../../../helpers';
 
 export default function Invoice(props: { startDate: Date, count: number, dataStatus: boolean }) {
     const [timeLimit] = useState(props.startDate);
 
-    const calculateTimeLeft = () => {
-        const difference = new Date(timeLimit).getTime() - new Date().getTime();
-
-        let timeLeft: any = {};
-
-        if (difference > 0) {
-            timeLeft = {
-                minutes: Math.floor((difference / 1000 / 60) % 60),
-                seconds: Math.floor((difference / 1000) % 60),
-            };
-        }
-
-        return timeLeft;
-    };
-
-
     const { main } = useStores();
-    const [timeLeft, setTimeLeft] = useState(calculateTimeLeft());
+    const [timeLeft, setTimeLeft] = useState(calculateTimeLeft(timeLimit));
 
     useEffect(() => {
         const invoiceTimeout = setTimeout(() => {
-            setTimeLeft(calculateTimeLeft());
+            setTimeLeft(calculateTimeLeft(timeLimit));
         }, 1000);
 
         if (props.count > 29) {
